Startup_project/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-environ==0.10.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
phonenumbers==8.13.18
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3